pyproject.toml
[tool.poetry]
name = "campo-conecta-back"
version = "0.1.0"
description = "Backend da solução do desafio 3 do hackathon Jornada Caatinga+ 2024"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.2"}
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
